[BE] TEST: comment test 분기 및 create 예외 케이스 추가
