+++ server/test/task.api.test.js
@@ -367,101 +367,3 @@ describe('delete task', () => {
     }
   });
 });
-
-describe('get comments', () => {
-  it('get comments 일반', async done => {
-    // given
-    const expectedCommentId = seeder.comments[0].id;
-    const taskId = seeder.tasks[1].id;
-
-    try {
-      // when
-      const res = await request(app)
-        .get(`/api/task/${taskId}/comment`)
-        .set('Authorization', `Bearer ${createJWT(seeder.users[0])}`);
-      const firstCommentId = res.body[0].id;
-
-      // then
-      expect(firstCommentId).toEqual(expectedCommentId);
-      done();
-    } catch (err) {
-      done(err);
-    }
-  });
-});
-
-describe('create comment', () => {
-  it('create comment 일반', async done => {
-    // given
-    const requestBody = {
-      content: '새로운 댓글',
-    };
-    const taskId = seeder.tasks[1].id;
-
-    try {
-      // when
-      const res = await request(app)
-        .post(`/api/task/${taskId}/comment`)
-        .set('Authorization', `Bearer ${createJWT(seeder.users[0])}`)
-        .send(requestBody);
-
-      // then
-      expect(res.status).toBe(status.SUCCESS.POST.CODE);
-      expect(res.body.message).toBe(status.SUCCESS.MSG);
-      done();
-    } catch (err) {
-      done(err);
-    }
-  });
-});
-
-describe('update comment', () => {
-  it('update comment 일반', async done => {
-    // given
-    const requestBody = {
-      content: '바뀐 댓글',
-    };
-    const taskId = seeder.tasks[1].id;
-    const commentId = seeder.comments[0].id;
-    try {
-      // when
-      const res = await request(app)
-        .put(`/api/task/${taskId}/comment/${commentId}`)
-        .set('Authorization', `Bearer ${createJWT(seeder.users[0])}`)
-        .send(requestBody);
-
-      // then
-      expect(res.status).toBe(status.SUCCESS.CODE);
-      expect(res.body.message).toBe(status.SUCCESS.MSG);
-      done();
-    } catch (err) {
-      done(err);
-    }
-  });
-});
-
-describe('delete comment', () => {
-  it('delete comment 일반', async done => {
-    // given
-    const requestBody = {
-      content: '바뀐 댓글',
-    };
-    const taskId = seeder.tasks[1].id;
-    const commentId = seeder.comments[0].id;
-
-    try {
-      // when
-      const res = await request(app)
-        .delete(`/api/task/${taskId}/comment/${commentId}`)
-        .set('Authorization', `Bearer ${createJWT(seeder.users[0])}`)
-        .send(requestBody);
-
-      // then
-      expect(res.status).toBe(status.SUCCESS.CODE);
-      expect(res.body.message).toBe(status.SUCCESS.MSG);
-      done();
-    } catch (err) {
-      done(err);
-    }
-  });
-});
